+++ common/src/main/java/com/mcvoice/ttvoice/plasmo/McvoicePvAddon.java
@@ -10,7 +10,7 @@ import su.plo.voice.api.server.PlasmoVoiceServer;
     id = "mcvoice",
     name = "MCvoice",
     scope = AddonLoaderScope.ANY_SERVER,
-    version = "0.2.4",
+    version = "0.2.5",
     authors = {"MCvoice"},
     dependencies = {}
 )